requirements.txt
streamlit
google-generativeai
groq
Pillow
PyPDF2
google-api-core
